Move BuildTxWith and associated types to a new module to avoid cyclical dependencies
